+++ app/src/main/java/com/mercadolibre/android/mlbusinesscomponentsapp/ButtonsActivity.java
@@ -20,14 +20,14 @@ public class ButtonsActivity extends AppCompatActivity {
                 .setTextInformation("Procesar Pago","Cargando")
                 .setColorText(R.color.ui_meli_red);
 
-        buttonProgress.setOnClickListener(v ->
-                Toast.makeText(ButtonsActivity.this,"hola", Toast.LENGTH_SHORT).show()
+        buttonProgress.setOnClickListener(v -> {
+                Toast.makeText(ButtonsActivity.this,"hola", Toast.LENGTH_SHORT).show();
+                buttonProgress.startAnimationCustom();
+            }
         );
 
-        new Handler().postDelayed(() -> buttonProgress.finishProgress(
-                R.color.ui_meli_red,
-                R.drawable.mercado_pago
-        ), 3000);
+        new Handler().postDelayed(buttonProgress::startAnimationCustom, 3000);
+        new Handler().postDelayed(buttonProgress::resetButton, 5000);
     }
 
     private OnFinishAnimationListener getFinishAnimationListener(){